in svg_image model in func create_svg added handling case when svg file is created and we only want to modify its height and width

+++ 5lab_django_intro/my_site/obrazkiApp/models.py
@@ -5,28 +5,42 @@ from django.contrib.auth.models import User
 import os
 from pathlib import Path
 
+base_dir = Path(__file__).resolve().parent.parent
+
 # Create your models here.
 class SVG_image(models.Model):
     name = models.CharField(max_length=100)
     width = models.PositiveIntegerField()
     height = models.PositiveIntegerField()
     permitted_users = models.ManyToManyField(User)
-    # svg = models.FileField(upload_to="images/")
 
     def __str__(self):
         return self.name
     
     def create_svg(self):
-        svg_content = f'<svg width="100" height="100"><rect width="{self.width}" height="{self.height}" /></svg>'
-        base_dir = Path(__file__).resolve().parent.parent
+
+        svg_header = f'<svg width="{self.width}" height="{self.height}" xmlns="http://www.w3.org/2000/svg">'
+        svg_end = '</svg>' 
+        svg_content = svg_header + svg_end
+
         file_path = base_dir / "media/images/" / (self.name + ".svg")
 
-        with open(file_path, 'w') as file:
-            file.write(svg_content)
+        if os.path.exists(file_path):
+            file_content = ""
+            with open(file_path, 'r') as file:
+                # otwieramy plik svg i czytamy cala zawartosc, nastepnie usuwamy
+                # naglowek svg i dodajemy nasz nowy naglowek
+                file_content = file.read()
+                svg_header_end_idx = file_content.find('>')
+                file_content = file_content[svg_header_end_idx + 1:]
+                file_content = svg_header + file_content
+            with open(file_path, 'w') as file:
+                file.write(file_content)
+        else:
+            with open(file_path, 'w') as file:
+                file.write(svg_content)
 
     def save(self, *args, **kwargs):
-        # Using the regular field, set the value of the read-only field.
         self.create_svg()
         # call the parent's save() method
         super(SVG_image, self).save(*args, **kwargs)
-    #     # return f'<svg width=100" height=100"><rect width="{self.width}" height="{self.height}" /></svg>'
